fix invoke

+++ src/main/java/com/example/EthanApiPlugin/EthanApiPlugin.java
@@ -367,9 +367,9 @@ public class EthanApiPlugin extends Plugin {
                               int var8) {
         Class invokeClass = client.getClass().getClassLoader().loadClass("ce");
         Method invoke = invokeClass.getDeclaredMethod("kc", int.class, int.class, int.class, int.class, int.class,
-                String.class, String.class, int.class, int.class, int.class);
+                String.class, String.class, int.class, int.class, byte.class);
         invoke.setAccessible(true);
-        invoke.invoke(null, var0, var1, var2, var3, var4, var5, var6, var7, var8, 14);
+        invoke.invoke(null, var0, var1, var2, var3, var4, var5, var6, var7, var8,(byte)14);
         invoke.setAccessible(false);
     }
